+++ backend/src/utils/file/config_reader.rs
@@ -21,6 +21,7 @@ use std::io::{self, Read};
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use tokio::fs;
+use tokio::sync::OnceCell;
 use shared::concat_string;
 use crate::utils::request::{is_uri};
 use url::Url;
@@ -308,6 +309,7 @@ pub async fn read_initial_app_config(
         custom_stream_response: Arc::new(ArcSwapAny::default()),
         access_token_secret: Default::default(),
         encrypt_secret: Default::default(),
+        ffprobe_available: Arc::new(OnceCell::new())
     };
     app_config.prepare(include_computed)?;
     //print_info(&app_config);